A bit of refactoring, help updates

+++ src/main/java/eu/jedrzej/jptest/input/Main.java
@@ -99,7 +99,7 @@ public class Main {
 
                     case "gbce":
                     case "index": {
-                        BigInteger index = GeometricalMeanCalculator.geometricalMean(StockHelper.getAllStocks());
+                        BigDecimal index = GeometricalMeanCalculator.geometricalMean(StockHelper.getAllStocks());
                         System.out.println("BGCE index: " + index);
                         break;
                     }
